Updated the OpenAPI specs to include document list endpoints (#366)
* Updated the OpenAPI specs to include document list endpoints

This PR also seems to include a boatload of other changes that have been introduced since the previous time it was exported. It is unknown what effects these might have on third parties, but the changes all seem to work when rendered in swagger-ui. No formal testing has been completed using these changes.

* Updated servers block to have the correct URL